pyproject.toml
[tool.poetry]
name = "jigsawhelper"
version = "0.1.0"
description = ""
authors = ["SpicyKitten <rathavilash@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
keyboard = "^0.13.5"
pillow = "^10.3.0"
pyqt6 = "^6.7.0"
screeninfo = "^0.8.1"

[tool.poetry.scripts]
server = "jigsawhelper.main:main"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
